+++ src/components/FileTree/FileFold.tsx
@@ -3,10 +3,8 @@ import { type FC, type PropsWithChildren, useMemo, useState } from "react";
 import { cn } from "../../util/cn";
 import { FileToolbar } from "./FileToolbar";
 import "./FileFolder.css";
-import type {
-    FileFolder,
-    FileKind,
-} from "../../features/Projects/stores/slices/files";
+import type { FileFolder } from "../../features/Projects/models/FileFolder";
+import type { FileKind } from "../../features/Projects/models/FileKind";
 import { useProject } from "../../features/Projects/stores/useProject";
 import { FileEntry, logoByKind } from "./FileEntry";
