+++ src/components/FileTree/FileToolbar.tsx
@@ -1,9 +1,7 @@
 import { assets } from "@kaplayjs/crew";
 import type { FC, PropsWithChildren } from "react";
-import {
-    type FileKind,
-    folderByKind,
-} from "../../features/Projects/stores/slices/files";
+import type { FileKind } from "../../features/Projects/models/FileKind";
+import { folderByKind } from "../../features/Projects/stores/slices/files";
 import { useProject } from "../../features/Projects/stores/useProject";
 
 type Props = PropsWithChildren<{
