+++ src/components/Toolbar/Projects.tsx
@@ -1,8 +1,8 @@
 import { assets } from "@kaplayjs/crew";
 import type { FC } from "react";
-import type { Asset } from "../../features/Projects/stores/slices/assets";
-import type { File } from "../../features/Projects/stores/slices/files";
-import type { Project } from "../../features/Projects/stores/slices/project";
+import type { Asset } from "../../features/Projects/models/Asset";
+import type { File } from "../../features/Projects/models/File";
+import type { Project } from "../../features/Projects/models/Project";
 import { useProject } from "../../features/Projects/stores/useProject";
 import { useEditor } from "../../hooks/useEditor";
 import { downloadBlob } from "../../util/download";
