+++ src/data/exampleList.json
@@ -831,7 +831,7 @@
         "category": "concepts",
         "group": "",
         "description": "How to use lerp to make simple tweenings",
-        "code": "kaplay({\n    background: [141, 183, 255],\n});\n\nloadHappy();\nloadBean();\n\nonLoad(() => {\n    // We'll create a bean with position, opacity and scale\n    const bean = add([\n        sprite(\"bean\"),\n        pos(center()),\n        anchor(\"center\"),\n        opacity(),\n        area(),\n        scale(),\n    ]);\n\n    // And a text to tell the user how to switch modes\n    const numberMode = add([\n        text(\"Press 1 or 2 to switch modes\", { font: \"happy\", size: 25 }),\n        scale(1),\n        pos(25),\n        anchor(\"left\"),\n        color(BLACK),\n    ]);\n\n    // Then we'll define the light and dark colors\n    const lightColor = rgb(141, 183, 255);\n    const darkColor = rgb(74, 48, 82);\n\n    let backgroundColor = lightColor; // Is the variable that defines the intended background color\n    let mode = 1; // Defines what \"mode\" we're in\n\n    // Lerp takes an \"initial value\", an \"ending value\" and a number from 0 to 1\n    // This number determines wheter it will return the initial value or the ending value\n    onUpdate(() => {\n        // This makes it so the background always follows the intended background color\n        // Creating an effect similar to a tweening, but much more simple\n        // The 0.5 will determine how \"fast\" or how closely the initial value will follow the ending value\n        // Being 0 not follow at all and 1 being instant\n        const newColor = lerp(getBackground(), backgroundColor, 0.5);\n        setBackground(newColor);\n\n        if (mode == 1) {\n            backgroundColor = lightColor;\n\n            // These lines make it so bean now follows the mouse position very slowly\n            // And the color of our text to turn from the current color to the dark color\n            bean.pos = lerp(bean.pos, mousePos(), 0.1);\n            bean.opacity = lerp(bean.opacity, 1, 0.5);\n            numberMode.color = lerp(numberMode.color, darkColor, 0.5);\n        }\n        else if (mode == 2) {\n            backgroundColor = darkColor;\n\n            // These lines make it so bean now follows center and makes it's scale always 1\n            bean.pos = lerp(bean.pos, center(), 0.5);\n            bean.scale = lerp(bean.scale, vec2(1), 0.5);\n            numberMode.color = lerp(numberMode.color, lightColor, 0.5);\n\n            // When bean is being hovered, now it's opacity will follow 1, highlighting it on hover\n            if (bean.isHovering()) {\n                bean.opacity = lerp(bean.opacity, 1, 0.1);\n                // If we press click, it's scale will go to 2, and since it's following a scale of 1, it wil go back to that\n                if (isMousePressed(\"left\")) {\n                    bean.scale = vec2(2);\n                }\n            }\n            // When bean is not being followed, it's opacity follows 0.1 very slowly\n            else {\n                bean.opacity = lerp(bean.opacity, 0.1, 0.1);\n            }\n        }\n\n        if (isKeyPressed(\"1\")) mode = 1;\n        else if (isKeyPressed(\"2\")) mode = 2;\n    });\n});",
+        "code": "kaplay({\n    background: [141, 183, 255],\n});\n\nloadHappy();\nloadBean();\n\nonLoad(() => {\n    // We'll create a bean with position, opacity and scale\n    const bean = add([\n        sprite(\"bean\"),\n        pos(center()),\n        anchor(\"center\"),\n        opacity(),\n        area(),\n        scale(),\n    ]);\n\n    // And a text to tell the user how to switch modes\n    const numberMode = add([\n        text(\"Press 1 or 2 to switch modes\", { font: \"happy\", size: 25 }),\n        scale(1),\n        pos(25),\n        anchor(\"left\"),\n        color(BLACK),\n    ]);\n\n    // Then we'll define the light and dark colors\n    const lightColor = rgb(141, 183, 255);\n    const darkColor = rgb(74, 48, 82);\n\n    let backgroundColor = lightColor; // Is the variable that defines the intended background color\n    let mode = 1; // Defines what \"mode\" we're in\n\n    // Lerp takes an \"initial value\", an \"ending value\" and a number from 0 to 1\n    // This number determines wheter it will return the initial value or the ending value\n    onUpdate(() => {\n        // This makes it so the background always follows the intended background color\n        // Creating an effect similar to a tweening, but much more simple\n        // The 0.5 will determine how \"fast\" or how closely the initial value will follow the ending value\n        // Being 0 not follow at all and 1 being instant\n        const newColor = lerp(getBackground(), backgroundColor, 0.5);\n        setBackground(newColor);\n\n        if (mode == 1) {\n            backgroundColor = lightColor;\n\n            // These lines make it so bean now follows the mouse position very slowly\n            // And the color of our text to turn from the current color to the dark color\n            bean.pos = lerp(bean.pos, mousePos(), 0.1);\n            bean.opacity = lerp(bean.opacity, 1, 0.5);\n            numberMode.color = lerp(numberMode.color, darkColor, 0.5);\n        }\n        else if (mode == 2) {\n            backgroundColor = darkColor;\n\n            // These lines make it so bean now follows center and makes its scale always 1\n            bean.pos = lerp(bean.pos, center(), 0.5);\n            bean.scale = lerp(bean.scale, vec2(1), 0.5);\n            numberMode.color = lerp(numberMode.color, lightColor, 0.5);\n\n            // When bean is being hovered, now its opacity will follow 1, highlighting it on hover\n            if (bean.isHovering()) {\n                bean.opacity = lerp(bean.opacity, 1, 0.1);\n                // If we press click, its scale will go to 2, and since it's following a scale of 1, it wil go back to that\n                if (isMousePressed(\"left\")) {\n                    bean.scale = vec2(2);\n                }\n            }\n            // When bean is not being followed, its opacity follows 0.1 very slowly\n            else {\n                bean.opacity = lerp(bean.opacity, 0.1, 0.1);\n            }\n        }\n\n        if (isKeyPressed(\"1\")) mode = 1;\n        else if (isKeyPressed(\"2\")) mode = 2;\n    });\n});",
         "difficulty": 0,
         "version": "master",
         "minVersion": "3001.0",
@@ -840,7 +840,7 @@
             "animation"
         ],
         "createdAt": "2025-04-27T12:23:46-05:00",
-        "updatedAt": "2025-04-27T12:23:46-05:00"
+        "updatedAt": "2025-04-27T12:50:27-05:00"
     },
     {
         "id": 45,
