+++ src/features/Projects/models/Asset.ts
@@ -1,0 +1,6 @@
+import type { UploadAsset } from "./UploadAsset";
+
+export type Asset = UploadAsset & {
+    // import function for kaplay
+    importFunction: string;
+};
