+++ src/features/Projects/models/AssetKind.ts
@@ -1,0 +1,1 @@
+export type AssetKind = "sprite" | "sound" | "font";
