+++ src/features/Projects/models/File.ts
@@ -1,0 +1,9 @@
+import type { FileKind } from "./FileKind";
+
+export type File = {
+    name: string;
+    path: string;
+    language: string;
+    value: string;
+    kind: FileKind;
+};
