+++ src/features/Projects/models/FileFolder.ts
@@ -1,0 +1,1 @@
+export type FileFolder = "root" | "scenes" | "assets" | "utils" | "objects";
