+++ src/features/Projects/models/FileKind.ts
@@ -1,0 +1,1 @@
+export type FileKind = "kaplay" | "main" | "scene" | "assets" | "util" | "obj";
