+++ src/features/Projects/models/ProjectMode.ts
@@ -1,0 +1,1 @@
+export type ProjectMode = "ex" | "pj";
